+++ CHANGELOG.rst
@@ -26,6 +26,18 @@ Latest versions (2.0.x)
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 
+Version 2.0.2 (2021-02-01)
+--------------------------
+- Added a new option to easily turn on/off interactive plotting: e.g., simply set ``cv.options.set(interactive=False)`` to turn off interactive plotting. This meta-option sets the other options ``show``, ``close``, and ``backend``.
+- Changed the logic of ``do_show``, such that ``do_show=False`` will never show a plot, even if ``cv.options.show`` is ``True``.
+- Added a new method, ``cv.diff_sims()``, that allows the differences in results between two simulations to be quickly calculated.
+- Removed the ``keys`` argument from ``cv.daily_stats()``, since non-default keys are had to validate.
+- Fixed a bug that prevented prognoses parameters from being correctly set to those from an earlier version.
+- Added an R usage example to the ``examples`` folder (matching the one in the FAQ).
+- Added additional tests, increasing test coverage from 72% to 88%.
+- *GitHub info*: PR `779 <https://github.com/amath-idm/covasim/pull/779>`__
+
+
 Version 2.0.1 (2021-01-31)
 --------------------------
 - Pinned ``xlrd`` version to 1.2.0 due to a bug introduced in the ``2.0.1`` version of ``xlrd`` (see `here <https://stackoverflow.com/questions/65250207/pandas-cannot-open-an-excel-xlsx-file>`__ for details).
@@ -149,7 +161,7 @@ Version 1.7.4 (2020-10-02)
 Version 1.7.3 (2020-09-30)
 --------------------------
 - Changed ``test_prob.apply()`` and ``test_num.apply()`` to return the indices of people that were tested
-- ``cvm.date(None)`` returns ``None`` instead of an empty list. Both ``cvm.date()`` and ``cvm.day()`` no longer raise errors if the list of inputs includes ``None`` entries
+- ``cvm.date(None)`` returns ``None`` instead of an empty list. Both ``cvm.date()`` and ``cvm.day()`` no longer raise errors if the list of inputs includes ``None`` entries.
 - *GitHub info*: PR `699 <https://github.com/amath-idm/covasim/pull/699>`__